# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RagInt Backend API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
...
    "/api/breakpoint": {
      "get": {
        "summary": "Get client breakpoint state",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "tour" }
          },
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Upsert client breakpoint state",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["state"],
                "properties": {
                  "kind": { "type": "string", "default": "tour" },
                  "client_id": { "type": "string" },
                  "state": { "type": "object" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Clear client breakpoint state",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "default": "tour" }
          },
          {
            "name": "X-Client-ID",
            "in": "header",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/tour/plan": {
      "post": {
        "summary": "Build a tour plan",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "zone": { "type": "string" },
                  "profile": { "type": "string" },
                  "duration_s": { "type": "integer" },
                  "stops_override": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/selling_points": {
      "get": {
        "summary": "List selling points for a stop",
        "parameters": [
          { "name": "stop_name", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Upsert a selling point",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["stop_name", "text"],
                "properties": {
                  "stop_name": { "type": "string" },
                  "text": { "type": "string" },
                  "weight": { "type": ["number", "string", "null"] },
                  "tags": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" } }
      },
      "delete": {
        "summary": "Delete a selling point",
        "parameters": [
          { "name": "stop_name", "in": "query", "required": false, "schema": { "type": "string" } },
          { "name": "text", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/api/selling_points/topn": {
      "get": {
        "summary": "Get TopN selling points",
        "parameters": [
          { "name": "stop_name", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "n", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "duration_s", "in": "query", "required": false, "schema": { "type": "integer" } },
          { "name": "profile", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/api/status": {
      "get": {
        "summary": "Request status + timings",
        "parameters": [
          { "name": "request_id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
...
  }
}